backend/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.9
psycopg2-binary==2.9.6
pydantic==1.10.7
alembic==1.10.3
python-dotenv==1.0.0
boto3==1.26.118
pytest==7.3.1
httpx==0.24.0
asyncpg==0.27.0
requests==2.28.2
email-validator==2.0.0
supabase==2.0.3
